Added second and first deliverable to backend and frontend
refactor in process.
Also added error html page to handle non responding server
