+++ .gitignore
@@ -1,4 +1,2 @@
 
 .DS_Store
-
-second_deliverable/*.csv
