+++ backend/controller/controller.go
@@ -1,9 +1,12 @@
 package controller
 
 import (
+	"encoding/csv"
+	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
+	"os"
 	"strconv"
 	"time"
 
@@ -15,6 +18,8 @@ import (
 
 // Requested errors
 var requestErrors []string
+
+const TechStackPath = "data/tech_stack.csv"
 
 // UseCase interface
 type UseCase interface {
@@ -35,6 +40,86 @@ func New(
 	r *render.Render,
 ) *MovieUseCase {
 	return &MovieUseCase{u, r}
+}
+
+// Transforms the object struct into a JSON string using the json.Marshal function
+func ConvertStructToJSON(object interface{}) string {
+	e, err := json.Marshal(object)
+	if err != nil {
+		return err.Error()
+	}
+	return string(e)
+}
+
+// Prints the error to the Log console and sends the message as a JSON response.
+func displayError(w http.ResponseWriter, message string) {
+	log.Println(message)
+	fmt.Fprintf(w, "%s", ConvertStructToJSON(model.Response{Title: "Error", Message: message}))
+
+}
+
+// Gathers the data lines from the csv file and parse it to a bi dimensional slice of strigs.
+func GetDataFromCSVFile(filePath string) [][]string {
+	csvFile, err := os.Open(filePath)
+	if err != nil {
+		fmt.Println("\n- An error ocurred while reading the file. \n", err)
+	} else {
+		fmt.Println("\nSuccessfully Opened CSV file")
+	}
+	csvLines, err := csv.NewReader(csvFile).ReadAll()
+	if err != nil {
+		log.Println("\n- An error ocurred while reading the file. \n", err)
+		return nil
+	}
+	return csvLines
+}
+
+// It process a bi dimensional array of strings into a slice of Item structs
+func ParseCSVDataToItemsList(csvLines [][]string) (listOfItems []model.Item) {
+	// Convert csv lines to a generic item structure and append them to the array of items
+	for _, line := range csvLines {
+		newItem := model.Item{
+			Id:    line[0],
+			Title: line[1],
+			Years: line[2],
+		}
+		listOfItems = append(listOfItems, newItem)
+		log.Println(newItem.Id + " " + newItem.Title + " ")
+	}
+	return
+}
+
+func (t *MovieUseCase) GetLanguages(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	csvLines := GetDataFromCSVFile(TechStackPath)
+	listOfItems := ParseCSVDataToItemsList(csvLines)
+	fmt.Fprintf(w, "%s", ConvertStructToJSON(listOfItems))
+}
+
+func (t *MovieUseCase) GetLanguageById(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	csvLines := GetDataFromCSVFile(TechStackPath)
+	listOfItems := ParseCSVDataToItemsList(csvLines)
+	// Obtain the query param id number from URL
+	keys, ok := r.URL.Query()["id"]
+	if !ok || len(keys[0]) < 1 {
+		displayError(w, "Url Param 'id' is missing!")
+		return
+	}
+	// Casting the string number to an integer
+	id, err := strconv.Atoi(keys[0])
+	if err != nil {
+		displayError(w, "The Id provided is wrong, please check it!")
+		return
+	}
+	// Validations: number is positive and that exists as index in the slice
+	if id >= len(listOfItems) || id < 0 {
+		displayError(w, "The Id doesn't seem to exist!")
+		return
+	}
+	// Get the object from slice using the id as index
+	obj := listOfItems[id]
+	fmt.Fprintf(w, "%s", ConvertStructToJSON(obj))
 }
 
 func (t *MovieUseCase) GetMovies(w http.ResponseWriter, r *http.Request) {
