+++ backend/data/tech_stack.csv
@@ -1,0 +1,8 @@
+0,javascript,6
+1,python,4
+2,c++,1
+3,go,0
+4,swift,1
+5,react,3
+6,css,4
+7,html,5
