--- deliverables/first_deliverable/data.csv
@@ -1,6 +1,0 @@
-0,javascript,6
-1,python,4
-2,c++,1
-3,go,0
-4,swift,1
-5,react,3
