--- deliverables/second_deliverable/README
@@ -1,16 +1,0 @@
-This project is using the API developed as first_deliverable, in order to run this second_deliverable we need to run the first_deliverable before:
-1. First you need to go to academy-go-q12021/first_deliverable/
-2. Compile with go build 
-3. execute the binary with ./academy-go-q12021
-
-Then we are ready to compile and run our second second_deliverable
-1. Make sure that the API from first_deliverable is running on port 8080.
-2. First you need to navigate to academy-go-q12021/second_deliverable/
-3. Compile with go build 
-4. execute the binary with ./academy-go-q12021
-5. Go to http://localhost:3000 and watch the list. 
-6. Click on any item to navigate.
-
-To test you need to:
-1. Make sure that the API from first_deliverable is running on port 8080.
-2. execute test with go test command
