--- deliverables/second_deliverable/main_test.go
@@ -1,84 +1,0 @@
-package main
-
-import (
-	"encoding/csv"
-	"fmt"
-	"log"
-	"os"
-	"testing"
-)
-
-var Items []Item = []Item{
-	{Id: "1", Title: "Netflix", Years: "5"},
-	{Id: "2", Title: "Dispney+", Years: "1"},
-	{Id: "3", Title: "HBO Max", Years: "4"},
-	{Id: "4", Title: "Paramount+", Years: "2"},
-	{Id: "5", Title: "Universal+", Years: "4"},
-}
-
-func GetDataFromCSVFile(filePath string) ([][] string)  {
-	csvFile, err := os.Open(filePath)
-	if err != nil {
-		fmt.Println("\n- An error ocurred while reading the file. \n", err)
-	} else {
-		fmt.Println("\nSuccessfully Opened CSV file") 
-	}
-	csvLines, err := csv.NewReader(csvFile).ReadAll()
-    if err != nil {
-		log.Println("\n- An error ocurred while reading the file. \n", err)
-		return nil
-    }
-	return csvLines
-}
-
-func ParseCSVDataToItemsList(csvLines [][]string) (listOfItems []Item ) { 
-	// Convert csv lines to a generic item structure and append them to the array of items
-    for _, line := range csvLines {
-		newItem := Item{
-            Id: line[0],
-            Title: line[1],
-			Years: line[2],
-        }
-        listOfItems = append(listOfItems, newItem)
-        log.Println(newItem.Id + " " + newItem.Title + " " + newItem.Years + " " )
-    }
-	return 
-}
-
-var csvFileName string = "test.csv"
-
-func TestWriteDataToCSVFile(t *testing.T) {
-	type args struct {
-		fileName string
-		items    []Item
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "Write struct Item to csv",
-			args: args{
-				fileName: csvFileName, 
-				items: Items,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// write the data
-			WriteDataToCSVFile(tt.args.fileName, tt.args.items)
-			// check if data was written properly
-			csvLines := GetDataFromCSVFile(tt.args.fileName)
-			listOfItems :=  ParseCSVDataToItemsList(csvLines)   
-			log.Print(listOfItems)
-			if len(listOfItems) != len(Items) { // is the same length
-				t.Errorf("Expected %v items but instead got %v!", len(Items), len(listOfItems))
-			}
-			if listOfItems[0].Id != Items[0].Id || listOfItems[0].Title != Items[0].Title { // has the same content
-				t.Errorf("Expected hardcoded item %v to match stored item on the csv %v!", Items[0], listOfItems[0])
-			}
-
-		})
-	}
-}
