--- deliverables/second_deliverable/result.csv
@@ -1,5 +1,0 @@
-tt0000009,Miss Jerry,Miss Jerry,1894,1894-10-09,Romance,45,USA,None,Alexander Black,Alexander Black,Alexander Black Photoplays,"""Blanche Bayliss, William Courtenay, Chauncey Depew""",The adventures of a female reporter in the 1890s.,5.9,154,,,,,1.0,2.0,https://m.media-amazon.com/images/M/MV5BNjA5Y2ZhYzctNDc1Yy00OGViLWI3NGUtOTYwZmE3NDFiYmIxXkEyXkFqcGdeQXVyNTIzOTk5ODM@._V1_SX300.jpg
-tt0000574,The Story of the Kelly Gang,The Story of the Kelly Gang,1906,1906-12-26,"""Biography, Crime, Drama""",70,Australia,None,Charles Tait,Charles Tait,J. and N. Tait,"""Elizabeth Tait, John Tait, Norman Campbell, Bella Cola, Will Coyne, Sam Crewes, Jack Ennis, John Forde, Vera Linden, Mr. Marshall, Mr. McKenzie, Frank Mills, Ollie Wilson""",True story of notorious Australian outlaw Ned Kelly (1855-80).,6.1,589,$ 2250,,,,7.0,7.0,https://m.media-amazon.com/images/M/MV5BNTY4ZDk5MzYtNjk2Zi00ZWY3LTgwZjUtNDc5MWEzMWFlOTQzXkEyXkFqcGdeQXVyNjU1MTEwMjI@._V1_SX300.jpg
-tt0001892,Den sorte drøm,Den sorte drøm,1911,1911-08-19,Drama,53,"""Germany, Denmark""",,Urban Gad,"""Urban Gad, Gebhard Schätzler-Perasini""",Fotorama,"""Asta Nielsen, Valdemar Psilander, Gunnar Helsengreen, Emil Albes, Hugo Flink, Mary Hagen""","""Two men of high rank are both wooing the beautiful and famous equestrian acrobat Stella. While Stella ignores the jeweler Hirsch, she accepts Count von Waldberg's offer to follow her home, ...""",5.8,188,,,,,5.0,2.0,https://m.media-amazon.com/images/M/MV5BNWVkNjUyYjUtZGU1Yi00YTM4LTg5ZGEtYmNlNDAzNWRjODM2XkEyXkFqcGdeQXVyMzE0MjY5ODA@._V1_SX300.jpg
-tt0002101,Cleopatra,Cleopatra,1912,1912-11-13,"""Drama, History""",100,USA,English,Charles L. Gaskill,Victorien Sardou,Helen Gardner Picture Players,"""Helen Gardner, Pearl Sindelar, Miss Fielding, Miss Robson, Helene Costello, Charles Sindelar, Mr. Howard, James R. Waite, Mr. Osborne, Harry Knowles, Mr. Paul, Mr. Brady, Mr. Corker""",The fabled queen of Egypt's affair with Roman general Marc Antony is ultimately disastrous for both of them.,5.2,446,$ 45000,,,,25.0,3.0,https://m.media-amazon.com/images/M/MV5BZTcyMGU4ZmYtYTVmYS00NWFkLWExM2MtYWRlYTlhOWYxNjYwXkEyXkFqcGdeQXVyMzE0MjY5ODA@._V1_SX300.jpg
-tt0002199,"""From the Manger to the Cross; or, Jesus of Nazareth""","""From the Manger to the Cross; or, Jesus of Nazareth""",1912,1913,"""Biography, Drama""",60,USA,English,Sidney Olcott,Gene Gauntier,Kalem Company,"""R. Henderson Bland, Percy Dyer, Gene Gauntier, Alice Hollister, Samuel Morgan, James D. Ainsley, Robert G. Vignola, George Kellog, J.P. McGowan""","""An account of the life of Jesus Christ, based on the books of the New Testament: After Jesus' birth is foretold to his parents, he is born in Bethlehem, and is visited by shepherds and wise...""",5.7,484,,,,,13.0,5.0,https://m.media-amazon.com/images/M/MV5BZTQ3MDdkMWEtYjJmMS00YmE3LTgxMDktOWUwZjczMTdhNGQ1XkEyXkFqcGdeQXVyODg3Mjk3Mjc@._V1_SX300.jpg
