--- frontend/html/item.html
@@ -1,196 +1,0 @@
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8" />
-    <meta
-        name="viewport"
-        content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <!-- Bootstrap CSS -->
-    <link
-        rel="stylesheet"
-        href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-        integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-        crossorigin="anonymous"
-    />    
-    <title>{{.Title}}</title>
-</head>
-<style>
-    .accent {
-        background-color: #f5c518;
-        border-color: #f5c518;
-    }
-    .primary {
-        background-color: #252525;
-    }
-    body {
-        background-color: rgb(251, 251, 251);
-        font-family: Roboto;
-    }
-    .main {
-        padding: 2rem;
-    }
-    .button-container {
-        margin: 1rem;
-    }
-    .wrapper {
-        color: whitesmoke;
-        display: flex;
-        flex-direction: column;
-    }
-    .image-container {
-        display: flex;
-        justify-content: center;
-        align-items: flex-start;
-        padding: 1rem;
-    }
-    .body-container {
-        display: flex;
-        flex-grow: 1;
-        flex-direction: column;
-    }
-    .poster {
-        /* height: fit-content;
-        min-width: fit-content;
-        max-height: 400px; */
-        background-color: rgb(11, 17, 35);
-    }
-    .primaryColor {
-        background-color: #151a9c;
-    }
-    .secondaryColor {
-        background-color: #3c42f3;
-    }
-    .terciaryColor {
-        background-color: #252525;
-    }
-    .accentColor {
-        background-color: #f5c518;
-        color: black !important;
-    }
-
-    .information {
-        border-radius: 1rem;
-        color: black;
-        padding: 3rem;
-    }
-    .pattern {
-        background: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAoAAAAKCAYAAACNMs+9AAAATklEQVQoU2O0O3X+PwMDA8MhM0NGEI0LMBKtEJsJD27YgW1R0DgEtwWrdUQrxGYLY8V5iDUdhghrKFOITfenmXFgW/jSF+H3DNEKsdkCAO99IAvXcD3VAAAAAElFTkSuQmCC)
-            repeat;
-    }
-    .header {
-        display: flex;
-        color: black;
-    }
-    .content {
-        display: flex;
-        flex-direction: column;
-        padding: 3rem;
-        margin-top: 1rem;
-    }
-    .shadow {
-        -webkit-box-shadow: 9px 10px 42px -3px rgba(0, 0, 0, 0.57);
-        -moz-box-shadow: 9px 10px 42px -3px rgba(0, 0, 0, 0.57);
-        box-shadow: 9px 10px 42px -3px rgba(0, 0, 0, 0.57);
-    }
-    .badge {
-        margin-top: 4px;
-        margin-bottom: 2px;
-        font-size: medium;
-        color: white;
-    }
-    .subsection {
-        display: flex;
-        flex-direction: column;
-    }
-</style>
-<div class="container-fluid main">
-    <div class="button-container">
-        <a class="btn btn-primary accent" href="/getMovies" role="button"
-            >Go Back</a
-        >
-    </div>
-    {{if .ImdbTitleId}}
-    <div class="wrapper container">
-        <div class="row header">
-            <div class="col-12 col-md-6 col-lg-4 image-container">
-                {{if .Poster}}
-                <img class="poster shadow" src="{{.Poster}}" onerror="this.onerror=null;this.src='https://i.pinimg.com/474x/d3/e4/8e/d3e48efc9fb9830d437972cc2f0514b2.jpg'"/>
-                {{else}}
-                <img
-                    class="poster shadow"
-                    src="https://i.pinimg.com/474x/d3/e4/8e/d3e48efc9fb9830d437972cc2f0514b2.jpg"
-                />
-                {{end}}
-            </div>
-            <div class="col-12 col-md-6 col-lg-8 container information shadow">
-                <!-- Most important information -->
-                <h1 class="display-4">{{.OriginalTitle}}</h1>
-                <h6 class="card-text">ID: {{.ImdbTitleId}}</h6>
-                <p class="lead">{{.Description}}</p>
-                <p class="card-text">Actors: {{.Actors}}</p>
-
-                <div class="">
-                    <div class="row">
-                        <p class="col-6 card-text"><b>Genre:</b> {{.Genre}}</p>
-                        <p class="col-6 card-text">
-                            <b>Duration:</b>
-                            {{.Duration}}
-                        </p>
-                    </div>
-                    <div class="row">
-                        <p class="col-6 card-text"><b>Year:</b> {{.Year}}</p>
-                        <p class="col-6 card-text">
-                            <b>Director:</b> {{.Director}}
-                        </p>
-                    </div>
-                    <div class="row">
-                        <p class="col-6 card-text">
-                            <b>Country:</b> {{.Country}}
-                        </p>
-                        <p class="col-6 card-text">
-                            <b>Language:</b> {{.Language}}
-                        </p>
-                    </div>
-                    <div class="row">
-                        <p class="col-6 card-text">
-                            <b>Average Vote:</b> {{.AvgVote}}
-                        </p>
-                        <p class="col-6 card-text">
-                            <b>Date Published:</b> {{.DatePublished}}
-                        </p>
-                    </div>
-                </div>
-
-                <!-- Info with chips -->
-                <span class="badge badge-pill primaryColor"
-                    >Production: {{.ProductionCompany}}</span
-                >
-                <span class="badge badge-pill secondaryColor"
-                    >Writer: {{.Writer}}</span
-                >
-                <span class="badge badge-pill terciaryColor"
-                    >Budget: {{.Budget}}</span
-                >
-                <span class="badge badge-pill accentColor"
-                    >Votes: {{.Votes}}</span
-                >
-                <span class="badge badge-pill primaryColor"
-                    >USA Gross Income: {{.UsaGrossIncome}}</span
-                >
-                <span class="badge badge-pill secondaryColor"
-                    >World Wide Gross Income: {{.WorlwideGrossIncome}}</span
-                >
-                <span class="badge badge-pill accentColor"
-                    >Meta Score: {{.Metascore}}</span
-                >
-                <span class="badge badge-pill primaryColor"
-                    >WorlReviews From Users: {{.ReviewsFromUsers}}</span
-                >
-                <span class="badge badge-pill secondaryColor"
-                    >Reviews From Critics: {{.ReviewsFromCritics}}</span
-                >
-            </div>
-        </div>
-    </div>
-    {{end}}
-</div>
