+++ frontend/main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"bufio"
+	"encoding/csv"
 	"encoding/json"
 	"flag"
 	"fmt"
@@ -20,6 +22,108 @@ const (
 	ExitAbnormalErrorLoadingConfiguration = iota
 	ExitAbnormalErrorLoadingCSVFile
 )
+
+//
+func GetTechStackList() (items []model.TechStackItem) {
+	// Get the http reponse from api localhost:8080 (first_deliverable)
+	resp, err := http.Get("http://localhost:8080/getTechStack")
+	if err != nil {
+		log.Fatalf(err.Error())
+		items = []model.TechStackItem{{Title: "", Id: "", Years: ""}}
+		defer resp.Body.Close()
+		return
+	} else {
+		defer resp.Body.Close()
+		// Print the HTTP response status.
+		fmt.Println("\n\tResponse status:", resp.Status)
+
+		// Print the first 5 lines of the response body.
+		scanner := bufio.NewScanner(resp.Body)
+		for i := 0; scanner.Scan() && i < 5; i++ {
+			json.Unmarshal([]byte(scanner.Text()), &items) // items slice
+		}
+		if err := scanner.Err(); err != nil {
+			panic(err)
+		}
+		return items
+	}
+}
+
+func GetTechStackById(id string) (item model.TechStackItem) {
+	// Get the http reponse from api localhost:8080 (first_deliverable)
+	var url string = "http://localhost:8080/getTechStackById?id=" + id
+	resp, err := http.Get(url)
+	if err != nil {
+		panic(err)
+	}
+	defer resp.Body.Close()
+
+	// Print the HTTP response status.
+	fmt.Println("\n\tResponse status:", resp.Status)
+
+	// Print the first 5 lines of the response body.
+	scanner := bufio.NewScanner(resp.Body)
+	for i := 0; scanner.Scan() && i < 5; i++ {
+		json.Unmarshal([]byte(scanner.Text()), &item) // items slice
+	}
+	if err := scanner.Err(); err != nil {
+		panic(err)
+	}
+	return item
+}
+
+func RenderItem(w http.ResponseWriter, r *http.Request) {
+	keys, ok := r.URL.Query()["id"]
+	if !ok || len(keys[0]) < 1 {
+		errorMessage := "Url Param 'id' is missing"
+		log.Println(errorMessage)
+		fmt.Fprintf(w, "%s", errorMessage)
+		return
+	}
+	// Casting the string number to an integer
+	id := keys[0]
+	item := GetTechStackById(id)
+
+	tmpl := template.Must(template.ParseFiles("html/item.html"))
+
+	if err := tmpl.Execute(w, item); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+}
+
+func GetAllMovies(w http.ResponseWriter, r *http.Request) {
+	movies := GetMovies()
+	tmpl := template.Must(template.ParseFiles("html/movies.html"))
+	data := model.PageData{
+		PageTitle:     "My Tech Stack",
+		TechStackItem: techStackItem,
+	}
+	tmpl.Execute(w, data)
+	WriteDataToCSVFile("result.csv", items)
+}
+
+func WriteDataToCSVFile(fileName string, items []model.TechStackItem) {
+	log.Println("Data: ", items)
+
+	csvfile, err := os.Create(fileName)
+	if err != nil {
+		log.Fatalln("Error creating file csv", err)
+	}
+	var writter *csv.Writer = csv.NewWriter(csvfile)
+
+	for _, item := range items {
+		strSlice := []string{item.Id, item.Title, item.Years}
+		fmt.Println(strSlice)
+		writter.Write(strSlice)
+	}
+	// Write any buffered items data to the underlying writer (standard output).
+	writter.Flush()
+
+	if err := writter.Error(); err != nil {
+		log.Fatalln("error writing csv:", err)
+	}
+
+}
 
 func GetMovies(queryParams model.QueryParameters) (response model.Response_All) {
 	// Get the http reponse from api localhost:8080 backend
@@ -121,7 +225,7 @@ func RenderMovies(w http.ResponseWriter, r *http.Request) {
 
 	response := GetMovies(model.QueryParameters{Items: queryParams.Items, ItemPerWorkers: 1, Type: queryParams.Type})
 
-	tmpl := template.Must(template.ParseFiles("html/index.html"))
+	tmpl := template.Must(template.ParseFiles("html/movies.html"))
 	data := model.Page_AllMovies{
 		PageTitle: "Cine+",
 		Movies:    response.Data,
@@ -182,6 +286,11 @@ func main() {
 		os.Exit(ExitAbnormalErrorLoadingConfiguration)
 	}
 
+	// Second deliverable
+	http.HandleFunc("/getTechStack", GetAllItems)
+	http.HandleFunc("/getTechStackById", RenderItem)
+
+	// Third deliverable
 	http.HandleFunc("/getMovies", RenderMovies)
 	http.HandleFunc("/getMovieById", RenderMovieById)
 	fmt.Printf("Web app running succesfully on port [%s].", cfg.HTTPPort)
