+++ frontend/model/model.go
@@ -34,7 +34,7 @@ type Movie struct {
 	Poster              string `json:"poster"`
 }
 
-// MovieSummary struct (Model): meant to be used for a list of movies.
+// MovieSummary struct (Model): meant to be used for a list of movies with only the neccesary fields.
 type MovieSummary struct {
 	ImdbTitleId   string `json:"imdb_title_id"`
 	OriginalTitle string `json:"original_title"`
@@ -42,6 +42,7 @@ type MovieSummary struct {
 	Poster        string `json:"poster"`
 }
 
+// Response_All useful for parsing the response from server with a list of MovieSummary as Data (Third deliverable)
 type Response_All struct {
 	Title         string         `json:"title"`
 	Message       string         `json:"message"`
@@ -50,6 +51,8 @@ type Response_All struct {
 	Errors        []string       `json:"errors"`
 	ExecutionTime string         `json:"execution_time"`
 }
+
+// Response_Single useful for parsing the response from server with a single Movie as Data (Third deliverable)
 type Response_Single struct {
 	Title         string   `json:"title"`
 	Message       string   `json:"message"`
@@ -59,12 +62,27 @@ type Response_Single struct {
 	ExecutionTime string   `json:"execution_time"`
 }
 
+// Page_MovieDetails struct will render a page with a page title and a list (slice/array) of MovieSummary (Third deliverable)
 type Page_AllMovies struct {
 	PageTitle string
 	Movies    []MovieSummary
 }
 
+// Page_MovieDetails struct will render a page with a single movie item (Third deliverable)
 type Page_MovieDetails struct {
 	PageTitle string
 	Movie     Movie
+}
+
+// PageData struct will render a page with a single interface struct item (Second deliverable)
+type PageData struct {
+	PageTitle     string
+	TechStackItem TechStackItem
+}
+
+// model.TechStackItem struct meant for the tech stack items (Second deliverable)
+type TechStackItem struct {
+	Id    string `json:"id"`
+	Title string `json:"title"`
+	Years string `json:"years"`
 }
